FingerGAN/requirements.txt
torch
torchvision
matplotlib
streamlit
pillow